Added functions to superimpose frames and to determine if selected bond vector is present.
I have drafted functions to superimpose the frames of the trajectory to
the initial frame in the trajectory. This will help to ensure that the
computed s2 order parameters are accurate. I have also added a function
that seeks to determine if the selected bond vector is present. If it
is, it continues to compute the s2 order parameters. If it isn’t an
error is printed letting the user know to make a new selection.