Bond Vector Selection, Alignment, and Table
I was able to get the alignment and bond vector selection functions to
work properly. I also fixed my previous get_table function so that it
prints the table of resid, s2 order parameters, and bond vector
selections. These columns are also aligned so that the output is easier
to read.