Updated s2_calculator and contact_sum functions
These are the most recent version of the s2_calculator and contact_sum
functions. For the s2_calculator, -1 is now printed if residue
selection is not present. For the contact_sum, the alternative function
was found to be faster. Therefore, I am trying to make this function
more general. In addition, the Ci values computed from the first
contact_sum function append to a Ci_list which can now be printed along
with s2 order parameters in a table.